api/pyproject.toml
[project]
name = "sara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask-pydantic~=0.13.1",
    "flask-restx~=1.3.0",
    "flask-sqlalchemy~=3.1.1",
    "psycopg2>=2.9.10",
    "pydantic~=2.11.7",
    "pydantic-settings~=2.6.0",
    "flask-cors~=4.0.0",
    "flask-migrate~=4.0.7",
    "litellm>=1.0.0",
    "loguru>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
